header implementations moved to c files
