+++ src/platformio.ini
@@ -12,10 +12,12 @@ monitor_speed = 9600
 ; ************************************************************
 
 [env:EACHINE_TX801]
+src_filter = +<*.c> +<targets/Eachine_TX801/*.c>
 build_flags =
 	-D EACHINE_TX801
 
 ; [env:EACHINE_TX526]
+; src_filter = +<*.c> +<targets/Eachine_TX526/*.c>
 ; build_flags =
 ; 	-D EACHINE_TX526
