+++ src/src/openVTxEEPROM.h
@@ -1,6 +1,10 @@
+#pragma once
+
+#include <stdint.h>
+
 #define versionEEPROM 0
 
-bool updateEEPROM = false;
+extern uint8_t updateEEPROM;
 
 typedef enum
 {
@@ -10,7 +14,7 @@ typedef enum
 
 typedef struct
 {
-    uint16_t version;    
+    uint16_t version;
     vtxMode_e vtxMode;
     uint16_t currFreq;
     uint8_t channel;
@@ -23,40 +27,8 @@ typedef struct
     uint8_t unlocked;
 } openVTxEEPROM;
 
-openVTxEEPROM myEEPROM;
+extern openVTxEEPROM myEEPROM;
 
-void defaultEEPROM()
-{
-    myEEPROM.version = versionEEPROM;
-    myEEPROM.vtxMode = SMARTAUDIO;
-    myEEPROM.currFreq = 5800;
-    myEEPROM.channel = 27;
-    myEEPROM.freqMode = 0;
-    myEEPROM.pitmodeInRange = 0;
-    myEEPROM.pitmodeOutRange = 0;
-    myEEPROM.currPowermW = 25;
-    myEEPROM.currPowerdB = 14;
-    myEEPROM.currPowerIndex = 1;
-    myEEPROM.unlocked = 1;
-    
-    EEPROM_put(0, myEEPROM);
-}
-    
-void readEEPROM()
-{
-    EEPROM_get(0, myEEPROM);
-
-    if (myEEPROM.version != versionEEPROM)
-    {
-        defaultEEPROM();
-    }    
-}
-    
-void writeEEPROM()
-{
-    if (updateEEPROM)
-    {
-        EEPROM_put(0, myEEPROM);
-        updateEEPROM = false;
-    }    
-}
+void defaultEEPROM(void);
+void readEEPROM(void);
+void writeEEPROM(void);
