+++ src/src/rtc6705.h
@@ -1,4 +1,6 @@
-#include <Arduino.h>
+#pragma once
+
+#include <stdint.h>
 
 #define SynthesizerRegisterA 0x00
 #define SynthesizerRegisterB 0x01
@@ -13,77 +15,9 @@
 #define MIN_FREQ 5000
 #define MAX_FREQ 5999
 
-void spiPinSetup()
-{
-  pinMode(SPI_CLOCK, OUTPUT);
-  digitalWrite(SPI_CLOCK, LOW);
-  pinMode(SPI_DATA, OUTPUT);
-  digitalWrite(SPI_DATA, LOW);
-  pinMode(SPI_SS, OUTPUT);
-  digitalWrite(SPI_SS, HIGH);
-}
-
-void sendBits(long data)
-{
-  digitalWrite(SPI_SS, LOW);
-  delayMicroseconds(1);
-
-  for (uint8_t i = 0; i < 25; i++)
-  {
-    digitalWrite(SPI_CLOCK, LOW);
-    delayMicroseconds(1);
-    digitalWrite(SPI_DATA, data & 0x1);
-    delayMicroseconds(1);
-    digitalWrite(SPI_CLOCK, HIGH);
-    delayMicroseconds(1);
-
-    data >>= 1;
-  }
-
-  digitalWrite(SPI_CLOCK, LOW);
-  digitalWrite(SPI_DATA, LOW);
-  digitalWrite(SPI_SS, HIGH);
-  delayMicroseconds(1);
-}
-
-void rtc6705ResetState()
-{
-  long data = StateRegister | (1 << 4) | (0b0 << 5);
-
-  sendBits(data);
-}
-
-void rtc6705PowerAmpOn()
-{
-  long data = PredriverandPAControlRegister | (1 << 4) | (0b00000100111110111111 << 5);
-
-  sendBits(data);
-}
-
-void rtc6705PowerAmpOff()
-{
-  long data = PredriverandPAControlRegister | (1 << 4) | (0b00000000000000000000 << 5);
-
-  sendBits(data);
-}
-
-void rtc6705WriteFrequency(long newFreq)
-{
-  long freq = 1000 * newFreq;
-  freq /= 40;
-  long SYN_RF_N_REG = freq / 64;
-  long SYN_RF_A_REG = freq % 64;
-
-  long data = SynthesizerRegisterB | (1 << 4) | (SYN_RF_A_REG << 5) | (SYN_RF_N_REG << 12);
-
-  rtc6705PowerAmpOff();
-  setPowermW(0);
-
-  sendBits(data);
-
-  if (!pitMode)
-  {
-    rtc6705PowerAmpOn();
-    setPowerdB(myEEPROM.currPowerdB);
-  }
-}
+void spiPinSetup(void);
+void sendBits(uint32_t data);
+void rtc6705ResetState(void);
+void rtc6705PowerAmpOn(void);
+void rtc6705PowerAmpOff(void);
+void rtc6705WriteFrequency(uint32_t newFreq);
